Update visual.c
adding -1 to the sector coordinates doesn't work with struct xy having unsigned char members and is adding 255 instead. so visual scanning toward lower sector numbers results in scanning sectors like 257,260 that don't exist. setting ix,iy to char is enough for the 0..9 range and is working with negative offsets, too (because of overflow)