Create create_20
This is a bash script to read a quiz file and choose 20 questions from it at random. To use it, it's best to have a directory for the file of the selected questions. In addition, the permissions must be set correctly, and the index file in the quiz directory must be modified to point to the file of selected questions. This is useful because many of the quiz categories have many questions. No one wants to have 50 or 100 questions to answer.